Created Downloader module
- Created an abstract class to handle multiple states.
- Added a freezed state, to notify on the presentation layer when the downloader has been paused, init, complete or inProgress.
Created a Provider class using riverpod to have a cleaner ui and more segmented interaction between the logic and the presentation.
Created a core class that implements an internal StreamSubscription, this is used to stop the download or to cancel the same download, this class is used with Downloader, to have a 2 way of communication.